requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.3
django-allauth==0.61.1
django-cache-url==3.4.5
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
environs==11.0.0
idna==3.6
marshmallow==3.21.1
oauthlib==3.2.2
packaging==23.2
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
mutagen==1.47.0
pillow==10.2.0
django-allauth==0.61.1
google-api-python-client==2.122.0
google-auth-oauthlib==1.2.0
coverage==7.4.4
django-debug-toolbar==4.3.0
gunicorn==21.2.0